### 2016-07-20 #41 @whedon
Submission: PsychoPhysioPipeline
**Submitting author:** @sbogutzky (<a href="http://orcid.org/0000-0002-9279-6549">Simon Bogutzky</a>)
**Repository:** <a href="https://github.com/sbogutzky/PsychoPhysioPipeline" target ="_blank">https://github.com/sbogutzky/PsychoPhysioPipeline</a>
**Version:** v2.0.0
