### 2016-07-29 #41 @arfon
@openjournals/joss-editors - any suggestions for potential reviewers here?
