### 2016-07-29 #41 @davclark
I got here via accepting review for #40. A quick look suggests that more exposition is needed (explaining terms, clearly defining the broader utility and audience), and much of this would be common across #40 and this submission.

@arfon - are there specific guidelines on what should be submitted separately and what combined? These two submissions are clearly distinct code bases, but they are closely inter-linked and the requirements of JOSS for statement of need would seem to be highly overlapping.
