### 2016-07-30 #41 @arfon
> @arfon - are there specific guidelines on what should be submitted separately and what combined? These two submissions are clearly distinct code bases, but they are closely inter-linked and the requirements of JOSS for statement of need would seem to be highly overlapping.

Not really @davclark but this is a good question. @sbogutzky is there are reason why this and #40 were submitted separately?
