### 2016-07-30 #41 @sbogutzky
Yes, a have some reasons:

1. You can use both tools independently
2. The PPC will be used while mobile activities; The PPP will be used stationary
3. Different authorship in both projects
4. Programming language is different

@davclark - which terms need more exposition? The audience is the flow research community e. g. European Flow Researchers Network (EFRN)
