### 2016-07-31 #41 @davclark
Those seem like good reasons @sbogutzky - I'll start with #40, then, as I feel that's where I likely have a bit more specific expertise. Lots of folks could easily comment on an analysis pipeline.

> @davclark - which terms need more exposition? The audience is the flow research community e. g. European Flow Researchers Network (EFRN)

Again, I'll restrict to #40. Feel free to apply relevant comments there to this submission, of course! The clarification of audience is helpful - in both cases, you aren't writing to hacker / programmer types, so you should be clear about the skillsets required to use your tools! I'll repeat this in a more formal review on #40.
