### 2016-08-03 #41 @sbogutzky
@davclark - I will add some of your recommendation of #40 to this submission tomorrow, too.
