### 2016-08-04 #41 @sbogutzky
@davclark - Thank for your recommendations. I'm ready with my revision.
