### 2016-08-04 #41 @sbogutzky
Editors, please also note that change the title to "PsychoPhysioPipeline: A Processing and Analysis Pipeline for Psychophysiological Research"

It is possible to change that on the Paper page?
