### 2016-08-16 #41 @davclark
Well, I lied a little. But better late than never. Note that I'm looking at the updated paper.md in the repo.

First a few small suggestions, not necessary for acceptance, but suggested:

1. Many R users don't know how to build and install packages. I could install `flow` with `R CMD build flow` and `R CMD INSTALL flow`, but package dependencies are not automatic this way. So, you could instead install `devtools` and use their `build()` and `install()` functions, which *do* automatically install your dependencies. I'd document the latter approach.
1. More generally, why not put everything in the flow package proper? Your scripts could easily become functions.
2. "Source directory" is a strange name - it could mean where the source code is. I'd say "source data". Also, this is brittle - if you don't include a trailing '/', the system breaks.
3. Location of outputs is given in the output of the script - this could be given in the docs (the README). Outputs are self-documented, but no formal docs.
3. "Next" prompt in subset script is confusing. Why is it even there?
4. Docs could explain what Kubios HRV is / where it comes from, and what RR is.
5. Tests would be nice

### Things that need a little work

1. Explain that the subset script is a GUI selection tool? More critically, it is not clear to me how I would do the obvious thing - selecting along an x range. The zooming in clips the data along the y range.
2. There's no explicit description of community/support
3. The pipeline is a bit muddled - at first glance it seems like fairly basic R scripts. The real meat seems to be in the flow package, but this submission draws the reader's focus to the R scripts in `processing`, which I see more as examples of how to use the flow package.

  @arfon do you have any guidance around how submissions to JOSS should coordinate true libraries with supporting / example scripts?

4. I'm going to take the perspective of a lazy person here. Since this is git, I can compare my results to your outputs with a `git diff`. But then I see differences which I need to visually diff like so:

  ```
-4.3,5,3.25,1,5,7,5,5,2016-03-14 18:18:53,"Running",0,1754583,1839311,1,"Bogutzky","Simon","1984-02-23"
+4.3,5,3.25,1,5,7,5,5,2016-03-14 18:18:53,"running",0,1754583,1839311,1,"bogutzky","simon","1984-10-31"
```

  The issue there is that I didn't capitalize things, as I believed I needed to match the case of the directories (though apparently not). So, barring explicit tests, some instructions to perhaps delete the processed-data directory, and also about what explicitly to type prior to a git diff (which should result in no difference) would be good.

  Coincidentally, I randomly said you were born on Hallowe'en of 1984. I got the year right 😮

5. DOI for Hreljac2000? doi:10.1016/S0966-6362(00)00045-X
6. The only text / code for KubiosHRV is in the README. It's not clear how I do this.
7. In get-stride-data.R, I have no intuitions about an appropriate threshold for "not detection" (also, is this "knot" detection?). So, I said 5. This yields lots of points around 150 strides/minute. Sounds fast!
8. Also in get-stride-data.R I let the thing sit for a while. It was not very clear that the plot wanted interaction. I also have no idea what I'm doing. After that, there are 265 plots where I can click on things and I don't know why...

At this point, a bit over an hour had passed, and I had exceeded my  threshold for digging around.

As-is, I don't think this is up to JOSS standards. Overall, more context is required to understand what the scripts return to me and expect me to interact with. More explicit testing instructions (ideally automated) would go a long way.

It seems like a useful piece of software (and an interesting project), so I encourage @sbogutzky to make such changes. Or alternatively, identify a reviewer with more expertise in this domain (~ sports physiology).
