### 2016-08-17 #41 @sbogutzky
@davclark -- Thanks again, unfortunately I'm a little bit busy, because of the revision of my PhD Thesis. Maybe I can make some changes on the weekend.

Some remarks:

For 1.) The visualization at at the end is only for checking the segment and its data. The range of each segment of data based on the times of self-reports in self-report.csv
If you have a session with motion data and 4 self-reports at minute 10' 20' 30' and 40', you will get four segments of data.

For 3.) The real meat is the pipeline. In the flow-package are only utility functions.
The pipeline produces from self-reports, raw ECG and motion data psychological, physiological and locomotory indicator, which can be used to find correlations between these branches.

For 4.) What is the issue?

For 5.) The doi is right! http://www.gaitposture.com/article/S0966-6362(00)00045-X/abstract

For 7.) The pipeline use the x rotation in deg/s to detect the midswing event of the gait cycle (step)?
200 is a good value.

For 7 and 8.) Ok, there are more instructions needed. For you: You have 265 (maybe) wrong step detection and has to control it -- Remove wrong detections and add missing detection. For few misleading detections, you can use the cloud plot (plot before the control plots) to remove outliers.
