### 2016-08-17 #41 @davclark
JOSS seems pretty relaxed about the review process, and I'm not in any hurry. And, I'm confident you can get this in shape, as none of my concerns are core to your project. For now if you need to put something on a CV, it's fair to say you're "in revision".

To respond to your query about (4) - there should be a reliable way (i.e., the user shouldn't have to think too hard) to test the pipeline. The way it's set up currently, the results of running the pipeline on your example data still incorporate whatever random values I type in. So, my results can differ from the (presumably) correct outputs included in the test data. Usually when testing, "different" == "wrong".

I showed some examples where the results are "differerent" but not in a meaningful way. In this particular case, you could suggest the values to type... probably in general you could suggest values for all of the steps? Then, you'd be able to check that the system is working properly with a git diff, and checking if files have a more recent modification date. Still not great for automated testing, but I'd call that the bare "good enough" - and it could pave the way towards automated testing.
