### 2016-08-18 #41 @arfon
> @arfon do you have any guidance around how submissions to JOSS should coordinate true libraries with supporting / example scripts?

Thanks for the review @davclark. I'm afraid I don't fully understand your question 😕. If you're asking whether we expect libraries (or utility functions as @sbogutzky describes them here) to be exercised in the tests then ideally yes. Quoting from the http://joss.theoj.org/about#reviewer_guidelines

> Tests

> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

In addition we have some reviewer guidance about making sure there are examples of how to actually use the software:

> Example usage

> The authors should include examples of how to use the software (ideally to solve real-world analysis problems).

As I said earlier though, I'm not sure I've exactly understood your question @davclark.
