### 2016-08-18 #41 @davclark
@arfon - I am pretty sure I didn't ask clearly. The test question was a separate thing, and I think I understand that clearly enough, and have communicated as much in my comments.

Right now, this software submission contains an R library (flow), in addition to analysis scripts which @sbogutzky considers the primary contribution. This is awkward, but I wasn't going to force him to separate this into two repositories. I was asking for editorial advice only on this issue.
