### 2016-08-23 #41 @arfon
> Right now, this software submission contains an R library (flow), in addition to analysis scripts which @sbogutzky considers the primary contribution. This is awkward, but I wasn't going to force him to separate this into two repositories. I was asking for editorial advice only on this issue.

@davclark gotcha, thanks.

> The pipeline is a bit muddled - at first glance it seems like fairly basic R scripts. The real meat seems to be in the flow package, but this submission draws the reader's focus to the R scripts in processing, which I see more as examples of how to use the flow package.

As you point out in this comment, by combining these two components this makes it somewhat more challenging to provide clear documentation for the user and tests/examples to demonstrate how to use the tool. All that said, I don't think we should force the authors into separate repositories provided they can clearly document how to use the tool.
