+++ report/case-study-flow-while-running.Rmd
@@ -24,9 +24,6 @@ options(OutDec = ",")
 # Remove all variables
 rm(list = ls(all = T))
 
-# Set working directory
-setwd("~/psychophysiopipeline/report")
-
 # Load libraries
 library("flow")
 library("tikzDevice")
@@ -37,7 +34,7 @@ library("coin")
 library("stringr")
 
 # Set paths
-root.path <- "C:/Users/sbogutzky/Desktop/daten (lokal)/2013/"
+root.path <- "/Volumes/DOS/daten/2013/"
 
 # Set directories
 processed.data.directory <- "processed-data/"
@@ -225,6 +222,37 @@ data.1[, 3] <- factor(data.1[, 3], levels = c("30'", "45'", "60'"), labels = c("
   
   # Clean up
   rm(feature.table, table, i)
+```
+
+```{r box-and-whisker-plots, include=FALSE}
+
+# Setup tikz device
+tikz("./tikz/x-x-box-and-whisker-plots.tex", width = 5.22, height = 5.22)
+
+plotBoxAndWhiskerPlot  = function(x, y, col, ...) {
+  boxplot(y ~ x, outline = FALSE, boxwex = .75, col = rgb(186/255, 187/255, 194/255, .5), ...)
+  points(as.numeric(x), y, xaxt = "n", cex = 2, col = col, main = "", ...)
+  grid(nx = NA, ny = NULL, col = rgb(186/255, 187/255, 194/255))
+  box()
+}
+
+# Set graphic parameter
+par(mai = c(.5, .5, .2, .1), mgp = c(2.5, 1, 0), mfcol = c(3, 3))
+
+plotBoxAndWhiskerPlot(data.1$measurement, data.1$flow, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(3, 7), main = "Generalfaktor", ylab = "$[1, 7]$", pch = 16)
+plotBoxAndWhiskerPlot(data.1$measurement, data.1$fluency, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(3, 7), main = "Glatter Verlauf", ylab = "$[1, 7]$", pch = 16)
+plotBoxAndWhiskerPlot(data.1$measurement, data.1$absorption, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(3, 7), main = "Absorbiertheit", ylab = "$[1, 7]$", pch = 16)
+
+frame()
+plotBoxAndWhiskerPlot(data.1$measurement, data.1$mean.hr, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(160, 190), main = "Herzfrequenz", ylab = "$1/min$", pch = 16)
+plotBoxAndWhiskerPlot(data.1$measurement, data.1$rmssd, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(4, 12), main = "RMSSD", ylab = "$ms$", pch = 16)
+
+
+frame()
+plotBoxAndWhiskerPlot(data.1$measurement, data.1$mean.jerk.cost, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(14, 24), main = "Bewegungsaufwand", ylab = "$\\times 10^3 \\: m^2 \\cdot s^{-5}$", pch = 16)
+plotBoxAndWhiskerPlot(data.1$measurement, data.1$mean.nsei, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(0, .4), main = "Norm. Shan. Entr. Index", ylab = "$[0, 1]$", pch = 16)
+
+
 ```
 
 ```{r variables-measurement-wise, echo=FALSE, warning=FALSE, results="asis"}
