### 2016-08-31 9d8fc99
Abbildung 5.2 angepasst
