+++ report/case-study-flow-while-running.Rmd
@@ -69,7 +69,7 @@ stride.data <- read.csv(stride.data.path)
 ecg.data <- read.csv(ecg.data.file.path)
 
 # Setup tikz device
-tikz("./tikz/5-2-daten.tex", width = 7.91, height = 5.22)
+tikz("./tikz/5_2_daten.tex", width = 7.91, height = 5.22)
 
 par(mai = c(.8, .8, .1, .1), mgp = c(2.5, 1, 0), mfcol = c(2, 4))
 
@@ -79,32 +79,32 @@ n <- 200
 ecg.data.subset <- ecg.data[ecg.data[, 1] > stride.data[n, 1] & ecg.data[, 1] < stride.data[n + 1, 1], ]
 x.p <- 1:nrow(ecg.data.subset) / nrow(ecg.data.subset) * 100
 
-plot(x.p, ecg.data.subset[, 2], type = "l", xlab = "Zeit (\\% Doppelschritt)", ylab = "EKG RA-LL ($mV$)", ylim = c(-3, 3), xaxs = "i")
+plot(x.p, ecg.data.subset[, 2], type = "l", xlab = "Zeit (\\% Bewegungsablauf)", ylab = "EKG RA-LL ($mV$)", ylim = c(-3, 3), xaxs = "i")
 grid(col = rgb(186/255, 187/255, 194/255))
 box()
-plot(x.p, ecg.data.subset[, 3], type = "l", xlab = "Zeit (\\% Doppelschritt)", ylab = "EKG LA-LL ($mV$)", ylim = c(-3, 3), xaxs = "i")
+plot(x.p, ecg.data.subset[, 3], type = "l", xlab = "Zeit (\\% Bewegungsablauf)", ylab = "EKG LA-LL ($mV$)", ylim = c(-3, 3), xaxs = "i")
 grid(col = rgb(186/255, 187/255, 194/255))
 box()
 
 # Plot kinematic Data
 kinematic.data.subset <- kinematic.data[kinematic.data[, 1] > stride.data[n, 1] & kinematic.data[, 1] < stride.data[n + 1, 1], ]
 x.p <- 1:nrow(kinematic.data.subset) / nrow(kinematic.data.subset) * 100
-plot(x.p, -kinematic.data.subset[, 2], type = "l", xlab = "Zeit (\\% Doppelschritt)", ylab = "Beschleunigung X ($m \\cdot s^{-2}$)", ylim = c(-30, 30), xaxs = "i", col = rgb(155/255, 193/255, 54/255))
+plot(x.p, -kinematic.data.subset[, 2], type = "l", xlab = "Zeit (\\% Bewegungsablauf)", ylab = "Beschleunigung X ($m \\cdot s^{-2}$)", ylim = c(-30, 30), xaxs = "i", col = rgb(155/255, 193/255, 54/255))
 grid(col = rgb(186/255, 187/255, 194/255))
 box()
-plot(x.p, -kinematic.data.subset[, 5], type = "l", xlab = "Zeit (\\% Doppelschritt)", ylab = "Winkelgeschwindigkeit X ($deg \\cdot s^{-1}$)", ylim = c(-600, 600), xaxs = "i", col = rgb(155/255, 193/255, 54/255))
+plot(x.p, -kinematic.data.subset[, 5], type = "l", xlab = "Zeit (\\% Bewegungsablauf)", ylab = "Winkelgeschwindigkeit X ($deg \\cdot s^{-1}$)", ylim = c(-600, 600), xaxs = "i", col = rgb(155/255, 193/255, 54/255))
 grid(col = rgb(186/255, 187/255, 194/255))
 box()
[…]
