### 2016-09-02 5dcd605
Fünf-Minutenabschnitte getestet
