+++ report/code-snippets/get-hrv-features.R
@@ -1,25 +1,34 @@
 hrv.features <- data.frame()
-hrv.feature.names <- c("mean.rr","sdnn","mean.hr","sdhr","rmssd","nn50","pnn50","sdann","sdnni","hrvtri","tinn","vlfpeakfft","lfpeakfft","hfpeakfft","vlfpowfft","lfpowfft","hfpowfft","vlfpowprfft","lfpowprfft","hfpowprfft","lfpownufft","hfpownufft","totpowfft","lfhffft","vlfpeakar","lfpeakar","hfpeakar","vlfpowar","lfpowar","hfpowar","vlfpowprar","lfpowprar","hfpowprar","lfpownuar","hfpownuar","totpowar","lfhfar","edr","sd1","sd2","apen","sampen","d2","dfa1","dfa2")
+hrv.feature.names <- c("mean.hr","rmssd")
 for (i in 1:nrow(fss.features)) {
   
   session.start <- fss.features[i, 9]
   measurement <- fss.features[i, 14]
   
-  hrv.feature.path <- paste(root.path, feature.directory, activity.directory, user.directory, "kubios-hrv-features", "/", strftime(session.start, format="%Y-%m-%d--%H-%M-%S"), "/", "imu-rn42-bd38-", measurement, "_hrv.txt", sep = "")
+  kubios.hrv.data.file.path <- paste(root.path, processed.data.directory, activity.directory, user.directory, strftime(session.start, format="%Y-%m-%d--%H-%M-%S"), "/", "imu-rn42-bd38-", measurement, "_hrv.txt", sep = "")
   rm(measurement)
-  
-  if(file.exists(hrv.feature.path)) {
-    hrv.feature.row <- read.csv(hrv.feature.path, nrows = 1)
-    hrv.feature.row <- hrv.feature.row[, 2:46]
+  if(file.exists(kubios.hrv.data.file.path)) {
+    
+    # Load heart beat times
+    source("../processing/code-snippets/get-kubios-hrv-data.R")
+    
+    kubios.hrv.data <- kubios.hrv.data[max(kubios.hrv.data[, 1], na.rm = T) - min.before.end * 60 < kubios.hrv.data[, 1], ]
+    heart.beat.times <- c(kubios.hrv.data[1, 1] - kubios.hrv.data[1, 2], kubios.hrv.data[, 1])
+    
+    rr <- diff(heart.beat.times * 1000)
+    mean.hr <- mean(60000 / rr, na.rm = T)
+    rmssd <- sqrt(sum(diff(rr)^2, na.rm = T)/(length(rr)-1))
+    
+    hrv.feature.row <- data.frame(mean.hr, rmssd)
     names(hrv.feature.row) <- hrv.feature.names
     hrv.features <- rbind(hrv.features, hrv.feature.row)
   } 
   
   else {
-    hrv.feature.row <- data.frame(t(rep(NA, 45)))
+    hrv.feature.row <- data.frame(t(rep(NA, 2)))
     names(hrv.feature.row) <- hrv.feature.names
     hrv.features <- rbind(hrv.features, hrv.feature.row)
   }
-  rm(hrv.feature.path, hrv.feature.row)
+  rm(kubios.hrv.data.file.path, hrv.feature.row)
 }
 rm(hrv.feature.names, i, session.start)
