+++ report/case-study-flow-while-running.Rmd
@@ -197,19 +197,21 @@ fss.features <- read.csv(fss.feature.path)
 rm(fss.feature.path)
     
 # Get HRV features
+min.before.end <- 5
 source("code-snippets/get-hrv-features.R")
 
 # Get stride features
-min.before.end <- 12
+min.before.end <- 5
 kinematic.data.file.name <- "imu-rn42-3b70"
 source("code-snippets/get-stride-features.R")
     
 # Get jerk cost features
-min.before.end <- 12
+min.before.end <- 5
 kinematic.data.file.name <- "imu-rn42-3b70"
 source("code-snippets/get-jc-features.R")
     
 # Get cardio-locomotor phase synchronization features
+min.before.end <- 5
 source("code-snippets/get-cls-features.R") 
 
 # Set factors
@@ -218,13 +220,13 @@ measurements <- factor(fss.features[, 14], levels = c(1, 2, 3, 4), labels = c("1
 sessions <- factor(fss.features[, 9], levels = c("2013-10-03 17:55:54", "2013-10-10 17:41:56", "2013-10-17 18:07:11", "2013-10-24 17:32:14", "2013-10-31 18:31:19", "2013-11-07 17:34:06"), labels = c("03.10", "10.10", "17.10", "24.10", "31.10", "07.11"))
   
 # Create data frame
-data.1 <- rbind(data.1, data.frame(subject = subjects, session = sessions, measurement = measurements, fss.features[, c(1:8)], hrv.features[, c(2, 3, 5)], cls.features[, 1:2], mean.ms.interval = stride.features[, 1], mean.jerk.cost = jc.features[, 1]))
+data.1 <- rbind(data.1, data.frame(subject = subjects, session = sessions, measurement = measurements, fss.features[, c(1:8)], hrv.features[, 1:2], cls.features[, 1:2], mean.ms.interval = stride.features[, 1], mean.jerk.cost = jc.features[, 1]))
 rownames(data.1) <- seq(length = nrow(data.1))
 
 # Clean up
 rm(fss.features, hrv.features, cls.features, stride.features, jc.features, subjects, sessions, measurements)
 
-data.1[, 18] <- data.1[, 18] / 10^3
+data.1[, 17] <- data.1[, 17] / 10^3
 ```
 
 # FKS Reabilität
@@ -263,11 +265,6 @@ print(table, caption.placement = "top", sanitize.rownames.function = function(x)
 rm(data.table.flow, data.table.fluency, data.table.absorption, table)
 ```
 
-```{r remove-observations, include=FALSE}
-data.1 <- data.1[-c(1, 5, 9, 13, 17, 21), ]
-data.1[, 3] <- factor(data.1[, 3], levels = c("30'", "45'", "60'"), labels = c("30'", "45'", "60'"))
-```
-
 # Variablen
 
 ```{r variables-session-wise, echo=FALSE, warning=FALSE, results="asis"}
@@ -283,17 +280,17 @@ data.1[, 3] <- factor(data.1[, 3], levels = c("30'", "45'", "60'"), labels = c("
[…]
