### 2016-09-04 #41 @sbogutzky
I did a revision.

What I couldn't provide:
- One library for all code - or a separation of the code
- Automated test suite

I made the most changes in the documentation (README.md). I provide all input values for the provided example data and a little bit more explanation. Some output files will after all differ from the example data output, because of the user interaction.
