+++ report/case-study-flow-while-running.Rmd
@@ -224,8 +224,9 @@ data.1 <- rbind(data.1, data.frame(subject = subjects, session = sessions, measu
 rownames(data.1) <- seq(length = nrow(data.1))
 
 # Clean up
-rm(fss.features, hrv.features, cls.features, stride.features, jc.features, subjects, sessions, measurements)
+rm(fss.features, hrv.features, cls.features, stride.features, jc.features, subjects, sessions, measurements, kinematic.data.file.name, min.before.end)
 
+data.1[, 16] <- 60000 / data.1[, 16]
 data.1[, 17] <- data.1[, 17] / 10^3
 ```
 
@@ -265,32 +266,12 @@ print(table, caption.placement = "top", sanitize.rownames.function = function(x)
 rm(data.table.flow, data.table.fluency, data.table.absorption, table)
 ```
 
-# Variablen
-
-```{r variables-session-wise, echo=FALSE, warning=FALSE, results="asis"}
-    
-  # Create table
-  options(stringsAsFactors = FALSE)
-  feature.table <- data.frame()
-  for (i in c(4, 5, 6, 13, 14, 16, 18)) {
-    feature.table <- rbind(feature.table, c(paste(formatC(round(tapply(data.1[, i], data.1[, 2], mean, na.rm=TRUE), 2), format = "f", digits = 2), ";", formatC(round(tapply(data.1[, i], data.1[, 2], sd, na.rm=TRUE), 2), format = "f", digits = 2)),  paste(formatC(round(mean(data.1[, i], na.rm=TRUE), 2), format = "f", digits = 2), ";", formatC(round(sd(data.1[, i], na.rm=TRUE), 2), format = "f", digits = 2))))
-  }
-  
-  colnames(feature.table) <- c(levels(data.1[, 2]), "Gesamt")
-  rownames(feature.table) <- c("Generalfaktor", "Glatter Verlauf", "Absorbiertheit", "Herzfrequenz (BPM)", "RMSSD ($ms$)", "Norm. Shan. Entr. Index", "Bew. ($\\times 10^3 \\: m^2 \\cdot s^{-5}$)")
-  
-  # Print latex table
-  table                   <- xtable(feature.table, label = "tab:variablen_nach_sitzung_fallstudie_laufen", align = "lyyyyyyy", caption = "Variablen der Fallstudie zum Laufen: Arithmetisches Mittel $\\pm$ Standardabweichung zu den sechs Sitzungen [$N = 4$] -+- \\emph{Anmerkung}: Bew. = Bewegungsaufwand.") 
-  print(table, floating.environment="sidewaystable", caption.placement = "top", sanitize.rownames.function = function(x){x}, comment = FALSE, booktabs = TRUE)
-  
-  # Clean up
-  rm(feature.table, table, i)
-```
+# Merkmale
 
 ```{r variables-session-wise-box-and-whisker-plots, include=FALSE}
 
 # Setup tikz device
-#tikz("./tikz/x-x-box-and-whisker-plots.tex", width = 5.22, height = 5.22)
+tikz("./tikz/5_6_ubersicht_nach_laufen_erste_studie.tex", width = 5.22, height = 5.22)
 
 plotBoxAndWhiskerPlot  = function(x, y, col, ...) {
[…]
