+++ report/case-study-flow-while-running.Rmd
@@ -177,7 +177,7 @@ rm(ecg.data, ecg.data.subset, ecg.data.subset.minima, kinematic.data, kinematic.
 
 ```
 
-\begin{figure}[t]
+\begin{figure}[ht]
 	\input{./tikz/5_5_grundlage_der_klps}
 	\caption[Grundlage der Berechnung der kardio-lokomotorischen Phasensynchronisation]{Grundlage der Berechnung der kardio-lokomotorischen Phasensynchronisation}
 	\label{fig:5_5_grundlage_der_klps}
@@ -297,12 +297,12 @@ plotBoxAndWhiskerPlot(data.1$session, data.1$mean.jerk.cost, col = rgb(0/255, 15
 plotBoxAndWhiskerPlot(data.1$session, data.1$mean.jerk.cost, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(14, 24), main = "Doppelschrittfr.", ylab = "$1/min$", pch = 16)
 
 plotBoxAndWhiskerPlot(data.1$session, data.1$mean.hr, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(160, 190), main = "Herzfrequenz", ylab = "$1/min$", pch = 16)
-plotBoxAndWhiskerPlot(data.1$session, data.1$mean.nsei, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(0, .5), main = "Norm. Shan. Entr. In.", ylab = "$[0, 1]$", pch = 16)
+plotBoxAndWhiskerPlot(data.1$session, data.1$mean.nsei, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(0, .5), main = "N. Shan. Entr. Index", ylab = "$[0, 1]$", pch = 16)
 
 dev.off()
 ```
 
-\begin{figure}[t]
+\begin{figure}[ht]
 	\input{./tikz/5_6_ubersicht_nach_laufen_erste_studie}
 	\caption[Übersicht der expliziten und impliziten Merkmale nach Läufen der ersten Studie]{Übersicht der expliziten und impliziten Merkmale nach Läufen der ersten Studie [$N = 4$]}
 	\label{fig:5_6_ubersicht_nach_laufen_erste_studie}
@@ -351,11 +351,17 @@ plotBoxAndWhiskerPlot(data.1$measurement, data.1$mean.jerk.cost, col = rgb(0/255
 plotBoxAndWhiskerPlot(data.1$measurement, data.1$mean.jerk.cost, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(14, 24), main = "Doppelschrittfr.", ylab = "$1/min$", pch = 16)
 
 plotBoxAndWhiskerPlot(data.1$measurement, data.1$mean.hr, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(160, 190), main = "Herzfrequenz", ylab = "$1/min$", pch = 16)
-plotBoxAndWhiskerPlot(data.1$measurement, data.1$mean.nsei, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(0, .5), main = "Norm. Shan. Entr. In.", ylab = "$[0, 1]$", pch = 16)
+plotBoxAndWhiskerPlot(data.1$measurement, data.1$mean.nsei, col = rgb(0/255, 152/255, 199/255, .5), ylim = c(0, .5), main = "N. Shan. Entr. Index", ylab = "$[0, 1]$", pch = 16)
 
 dev.off()
 
 ```
+
+\begin{figure}[ht]
+	\input{./tikz/5_7_ubersicht_nach_messzeitpunkten_erste_studie}
[…]
