+++ report/case-study-flow-while-running.Rmd
@@ -62,7 +62,7 @@ ecg.data.file.name <- "imu-rn42-bd38-3.csv"
 kinematic.data.path <- paste(root.path, processed.data.directory, activity.directory, user.directory, date.directory, kinematic.data.file.name, sep = "")
 stride.data.path <- paste(root.path, processed.data.directory, activity.directory, user.directory, date.directory, stride.data.file.name, sep = "")
 ecg.data.file.path <- paste(root.path, processed.data.directory, activity.directory, user.directory, date.directory, ecg.data.file.name, sep = "")
-    
+
 # Load kinematic and stride data
 kinematic.data <- read.csv(kinematic.data.path)
 stride.data <- read.csv(stride.data.path)
@@ -115,9 +115,9 @@ rm(stride.data, ecg.data.subset, kinematic.data.subset, date.directory, ecg.data
 ```
 
 \begin{sidewaysfigure}
-	\input{./tikz/5_2_daten}
-	\caption[Aufgenommene Datenströme beim Laufen]{Aufgenommene Datenströme eines Bewegungsaublaufs in der ersten Studie zum Flow-Erleben beim Laufen. Von links nach rechts: (schwarz) EKG-Ableitung RA-LL und EKG-Ableitung LA-LL; (grün) Beschleunigung in X-Richtung und Winkelgeschwindigkeit um die X Achse; (blau) Beschleunigung in Y-Richtung und Winkelgeschwindigkeit um die Y-Achse; (rot) Beschleunigung in Z-Richtung und Winkelgeschwindigkeit um die Z-Achse}
-	\label{fig:5_2_daten}
+\input{./tikz/5_2_daten}
+\caption[Aufgenommene Datenströme beim Laufen]{Aufgenommene Datenströme eines Bewegungsaublaufs in der ersten Studie zum Flow-Erleben beim Laufen. Von links nach rechts: (schwarz) EKG-Ableitung RA-LL und EKG-Ableitung LA-LL; (grün) Beschleunigung in X-Richtung und Winkelgeschwindigkeit um die X Achse; (blau) Beschleunigung in Y-Richtung und Winkelgeschwindigkeit um die Y-Achse; (rot) Beschleunigung in Z-Richtung und Winkelgeschwindigkeit um die Z-Achse}
+\label{fig:5_2_daten}
 \end{sidewaysfigure}
 
 ```{r cls-basis, include=FALSE}
@@ -147,9 +147,9 @@ segments(x0 = ecg.data.subset$timestamp.ms[ecg.data.subset.minima] / 1000, y0 = 
 
 arrows(x0 = min(ecg.data.subset$timestamp.ms / 1000), y0 = -600, max(ecg.data.subset$timestamp.ms / 1000), length = 0.05)
 
-text(kinematic.data.subset$timestamp.ms[kinematic.data.subset.minima] / 1000, c(-800, -800),  as.character(round(kinematic.data.subset$timestamp.ms[kinematic.data.subset.minima] / 1000 - 2200, 2)))
+text(kinematic.data.subset$timestamp.ms[kinematic.data.subset.minima] / 1000, c(-800, -800), as.character(round(kinematic.data.subset$timestamp.ms[kinematic.data.subset.minima] / 1000 - 2200, 2)))
 
-text(ecg.data.subset$timestamp.ms[ecg.data.subset.minima] / 1000, c(-1000, -1000),  as.character(round(ecg.data.subset$timestamp.ms[ecg.data.subset.minima] / 1000 - 2200, 2)))
[…]
