+++ report/case-study-flow-while-walking.Rmd
@@ -1,7 +1,7 @@
 ---
-title: "Einzelfallanalyse Flow beim Gehen (intraindividell)"
+title: "Flow beim Gehen (intraindividell)"
 author: "Simon Bogutzky"
-date: "Juni 2016"
+date: "September 2016"
 header-includes:
   - \usepackage[utf8]{inputenc} 
   - \usepackage[T1]{fontenc}
@@ -25,19 +25,19 @@ options(OutDec = ",")
 rm(list = ls(all = T))
 
 # Set working directory
-setwd("~/psychophysiopipeline/report")
+setwd("~/Entwicklung/projects/bogutzky/repositories/psychophysiopipeline/report")
 
 # Load libraries
 library("flow")
 library("tikzDevice")
 library("xtable")
 library("apaTables")
-library("lme4") 
+library("lme4")
 library("coin")
 library("stringr")
 
 # Set paths
-root.path <- "C:/Users/sbogutzky/Desktop/daten (lokal)/2014/"
+root.path <- "/Volumes/DOS/daten/2014/"
 
 # Set directories
 processed.data.directory <- "processed-data/"
@@ -46,134 +46,64 @@ user.directory <- "grueter-barbara/"
 
 ```
 
-# Rohdaten
-
-```{r raw-data, include=FALSE}
-
-# Set directories
-date.directory <- "2014-05-27--12-34-52/"
-
-# Set file names
-kinematic.data.file.name <- "imu-rn42-3b70-3.csv"
-stride.data.file.name <- "imu-rn42-3b70-stride-data-3.csv"
-ecg.data.file.name <- "imu-rn42-bd38-3.csv"
-
-# Set paths
-kinematic.data.path <- paste(root.path, processed.data.directory, activity.directory, user.directory, date.directory, kinematic.data.file.name, sep = "")
-stride.data.path <- paste(root.path, processed.data.directory, activity.directory, user.directory, date.directory, stride.data.file.name, sep = "")
-ecg.data.file.path <- paste(root.path, processed.data.directory, activity.directory, user.directory, date.directory, ecg.data.file.name, sep = "")
-    
-# Load kinematic and stride data
-kinematic.data <- read.csv(kinematic.data.path)
-stride.data <- read.csv(stride.data.path)
-ecg.data <- read.csv(ecg.data.file.path)
-
-# Setup tikz device
-tikz("./tikz/x-x-daten.tex", width = 7.91, height = 5.22)
-
-par(mai = c(.8, .8, .1, .1), mgp = c(2.5, 1, 0), mfcol = c(2, 4))
-
-n <- 600
-
-# Plot ECG data
-ecg.data.subset <- ecg.data[ecg.data[, 1] > stride.data[n, 1] & ecg.data[, 1] < stride.data[n + 1, 1], ]
-x.p <- 1:nrow(ecg.data.subset) / nrow(ecg.data.subset) * 100
-
-plot(x.p, ecg.data.subset[, 2], type = "l", xlab = "Zeit (\\% Doppelschritt)", ylab = "EKG RA-LL ($mV$)", ylim = c(-3, 3), xaxs = "i")
-grid(col = rgb(186/255, 187/255, 194/255))
-box()
-plot(x.p, ecg.data.subset[, 3], type = "l", xlab = "Zeit (\\% Doppelschritt)", ylab = "EKG LA-LL ($mV$)", ylim = c(-3, 3), xaxs = "i")
-grid(col = rgb(186/255, 187/255, 194/255))
-box()
-
-# Plot kinematic Data
-kinematic.data.subset <- kinematic.data[kinematic.data[, 1] > stride.data[n, 1] & kinematic.data[, 1] < stride.data[n + 1, 1], ]
[…]
