+++ report/case-study-flow-while-running.Rmd
@@ -69,7 +69,7 @@ stride.data <- read.csv(stride.data.path)
 ecg.data <- read.csv(ecg.data.file.path)
 
 # Setup tikz device
-tikz("./tikz/5_2_daten.tex", width = 7.91, height = 5.22)
+tikz("./tikz/rohdaten_1.tex", width = 7.91, height = 5.22)
 
 par(mai = c(.8, .8, .1, .1), mgp = c(2.5, 1, 0), mfcol = c(2, 4))
 
@@ -115,14 +115,14 @@ rm(stride.data, ecg.data.subset, kinematic.data.subset, date.directory, ecg.data
 ```
 
 \begin{sidewaysfigure}
-\input{./tikz/5_2_daten}
+\input{./tikz/rohdaten_1}
 \caption[Aufgenommene Datenströme beim Laufen]{Aufgenommene Datenströme eines Bewegungsaublaufs in der ersten Studie zum Flow-Erleben beim Laufen. Von links nach rechts: (schwarz) EKG-Ableitung RA-LL und EKG-Ableitung LA-LL; (grün) Beschleunigung in X-Richtung und Winkelgeschwindigkeit um die X Achse; (blau) Beschleunigung in Y-Richtung und Winkelgeschwindigkeit um die Y-Achse; (rot) Beschleunigung in Z-Richtung und Winkelgeschwindigkeit um die Z-Achse}
-\label{fig:5_2_daten}
+\label{fig:rohdaten_1}
 \end{sidewaysfigure}
 
 ```{r cls-basis, include=FALSE}
 
-tikz("./tikz/5_5_grundlage_der_klps.tex", width = 5.22, height = 5.22)
+tikz("./tikz/grundlage_klps.tex", width = 5.22, height = 5.22)
 
 par(mai = c(.1, .1, .1, .1), mgp = c(.1, .1, 0), mfcol = c(1, 1))
 
@@ -177,10 +177,10 @@ rm(ecg.data, ecg.data.subset, ecg.data.subset.minima, kinematic.data, kinematic.
 
 ```
 
-\begin{figure}[ht]
-\input{./tikz/5_5_grundlage_der_klps}
+\begin{figure}[!htb]
+\input{./tikz/grundlage_klps}
 \caption[Grundlage der Berechnung der kardio-lokomotorischen Phasensynchronisation]{Grundlage der Berechnung der kardio-lokomotorischen Phasensynchronisation}
-\label{fig:5_5_grundlage_der_klps}
+\label{fig:grundlage_klps}
 \end{figure}
 
 ```{r load-data, include=FALSE}
@@ -251,15 +251,15 @@ rm(fss.measurements, fss.features)
 ```{r fss-reability-running-latex-tables, echo=FALSE, warning=FALSE, results="asis"}
 
 # Print latex table
-table <- xtable(data.table.flow, label = "tab:generalfaktor_fallstudie_laufen", align = "lrrr", caption = "Arithmetisches Mittel, Standardabweichung und Item-Faktor-Korrelation der Items des Generalfaktors der ersten Studie zum Flow-Erleben beim Laufen") 
+table <- xtable(data.table.flow, label = "tab:generalfaktor_1", align = "lrrr", caption = "Arithmetisches Mittel, Standardabweichung und Item-Faktor-Korrelation der Items des Generalfaktors der ersten Studie zum Flow-Erleben beim Laufen") 
 print(table, caption.placement = "top", sanitize.rownames.function = function(x){x}, comment = FALSE, booktabs = TRUE)
 
 # Print latex table
[…]
