### 2016-09-08 #41 @davclark
@sbogutzky, this is much easier for a clueless user like me to get started and know they are running the examples properly. I offer a few more comments below, but please attend in particular to "FIX" in particular.

Actual bug - figure saving dialog on mac doesn't work. It asks you to open an existing file. It should allow to specify a file, right? (should definitely FIX if you have access to a mac. If not, let me know, I'll see if I can figure it out).

I still find some of the visualizations a bit mystifying - for example, the plot after `subset-data.R` clearly includes some non-activity data, and is mostly a giant black blob (this has been zoomed a bit, resulting in a shift of the data to the right):

![image](https://cloud.githubusercontent.com/assets/120710/18348919/4585aaae-759c-11e6-985b-0f53c1fc7608.png)

Also, I think you mean "knot detection" not "not detection."

In my kicking the tires, `get-stride-data.R` didn't report when it had trouble finding the file (I entered wrong data). Not sure why that happened... (ideally FIX, but not necessary) Otherwise, this one makes more sense to me in terms of what's plotted. Nice stride, Simon! The final plot again starts as a solid blob.

Minor Type-o FIXes: "Worte" in `'get-jerk-cost-data.R`. Also - the README says this doesn't write anything...`get-cardiolocomotor-phase-sychronization.R`, 'optinal' -> 'optional'. (This plot is again very dense.)

The report is a bit unclear - what is the baseline referring to? Should I just use the running data? It would be nice to have a report that works with the example-data (minor FIX).
