### 2016-09-08 #41 @davclark
Hopefully it's clear from the above what should definitely be fixed - these things are very minor, and I offered to help with the mac issue. Please address those and anything else, and let me know. Should be quick!
