### 2016-09-08 #41 @sbogutzky
@davclark - First thanks for reviewing again!

Unfortunately, by the zoom and printing task I'm also only a user. I use for this task the R package [zoom](https://cran.r-project.org/web/packages/zoom/zoom.pdf). I also using a mac and have the same bugs. I cannot print from XQuartz and I can also go above the borders of my data.

I meant impossible detection or no detection (fixed)

Worte (fixed); the README said processing/compute-optimal-cut-off-frequencies.R has no output

optinal (fixed)

In the report I load now the example data. The report is for analysis purposes. It is only a template.
Now you have only one subject with one measurement. In real life you have many participant and measurements. Than you can conduct statistical analysis like correlation, regression, ANOVA, effect test, and so on. But that is up to the user. The pipeline only prepares the data.

I guess, I'm ready
