### 2016-09-08 #41 @davclark
As I understand things, this is "good enough" for JOSS. It still lacks automated tests and clear community guidelines, though these don't appear to be necessary for publication per reviewer guidelines.

@arfon let me know if you're ready to proceed on that basis, or if further input is needed!
