+++ processing/get-cardiolocomotor-phase-sychronization.R
@@ -26,15 +26,19 @@ library(flow)
 library(zoom)
 
 # User input
-root.directory.path <- readline("Quellverzeichnis > ")  
-first.name <- readline("Vorname der Untersuchungsperson > ")
-last.name <- readline("Nachname der Untersuchungsperson > ")
-activity <-  readline("Aktivität der Untersuchung > ")
-ecg.data.file.name <- readline("Dateiname der Datei mit EKG-Daten (ohne .csv) > ")
-kinematic.data.file.name.1 <- readline("Dateiname der ersten Datei mit kinematischen Daten (ohne .csv) > ")
-kinematic.data.file.name.2 <- readline("Dateiname der zweiten Datei mit kinematischen Daten (ohne .csv / optimal) > ")
-time.window.s <- as.numeric(readline("Zeitfenster für die Index-Berechnung  (s) > "))
+root.directory.path <- "/Volumes/DOS/daten/2016/" # readline("Quellverzeichnis > ")  
+#first.name <- readline("Vorname der Untersuchungsperson > ")
+#last.name <- readline("Nachname der Untersuchungsperson > ")
+#activity <-  readline("Aktivität der Untersuchung > ")
+ecg.data.file.name <- "imu-rn42-bd38" #readline("Dateiname der Datei mit EKG-Daten (ohne .csv) > ")
+kinematic.data.file.name.1 <- "imu-rn42-3b70" #readline("Dateiname der ersten Datei mit kinematischen Daten (ohne .csv) > ")
+kinematic.data.file.name.2 <- "" # "imu-rn42-bc98" #readline("Dateiname der zweiten Datei mit kinematischen Daten (ohne .csv / optimal) > ")
+time.window.s <- 15 #as.numeric(readline("Zeitfenster für die Index-Berechnung  (s) > "))
 if(is.na(time.window.s)) time.window.s = 30
+
+user.directories <- paste(list.dirs("/Volumes/DOS/daten/2016/processed-data/running", recursive = F, full.names = F), "/", sep = "")
+
+for(user.directory in user.directories) {
 
 # Set directory paths
 source("./code-snippets/set-directory-paths.R")
@@ -73,6 +77,8 @@ for (self.report.file.name in self.report.file.names) {
       stride.data.1 <- stride.data.1[order(stride.data.1[, 1]),]
       stride.times <- stride.data.1[, 1] / 1000
       
+      summary(diff(stride.data.1[, 1]))
+      
       # Load heart beat times
       source("./code-snippets/get-kubios-hrv-data.R")
       heart.beat.times <- c(kubios.hrv.data[1, 1] - kubios.hrv.data[1, 2], kubios.hrv.data[, 1])
@@ -84,7 +90,7 @@ for (self.report.file.name in self.report.file.names) {
       y.lim <- c(min(mean(spm, na.rm = TRUE) - sd(spm, na.rm = TRUE) * 2, mean(bpm, na.rm = TRUE) - sd(bpm, na.rm = TRUE) * 2), max(mean(spm, na.rm = TRUE) + sd(spm, na.rm = TRUE) * 2, mean(bpm, na.rm = TRUE) + sd(bpm, na.rm = TRUE) * 2))
       
       # Plot
-      par(mfcol = c(3, 1), mar = c(3.5, 4, 2, 4) + 0.1, mgp = c(2.5, 1, 0))
+      par(mfcol = c(4, 1), mar = c(3.5, 4, 2, 4) + 0.1, mgp = c(2.5, 1, 0))
       
       # SPM vs. BPM 
       plot(stride.times, spm, xlab = "", ylab = "Mittlerer Schritt & Mittlerer HR", xaxt = "n", xlim = time.range.s, ylim = y.lim, xaxs = "i", pch = 23, bg = rgb(0/255, 152/255, 199/255))
[…]
