### 2016-09-09 #41 @arfon
> As I understand things, this is "good enough" for JOSS. It still lacks automated tests and clear community guidelines, though these don't appear to be necessary for publication per reviewer guidelines.

Thanks for the followup @davclark. Our reviewer guidelines are pretty clear on tests but less so on community guidelines.

On tests we say:

> Good: An automated test suite hooked up to an external service such as Travis-CI or similar
> OK: Documented manual steps that can be followed to check the expected functionality of the software (e.g. a sample input file to assert behaviour)
> Bad (not acceptable): No way for you the reviewer to check whether the software works

Looking at https://github.com/sbogutzky/PsychoPhysioPipeline#processing-steps it seems like we're coming in under the 'OK' category here - there is some example data that can be used to verify the functionality.

On community guidelines we say:

> There should be clear guidelines for third-parties wishing to:

> - Contribute to the software
> - Report issues or problems with the software
> - Seek support

https://github.com/sbogutzky/PsychoPhysioPipeline#author-and-contribution says:

> The authors (Simon Bogutzky) welcome external contributors to freely use and extend this software. If you need some help, please write an issue.

I think this is probably the minimal version of what is required to comply with this requirement.

So yes, I think we're good to move to an accept decision on this paper. @sbogutzky at this point could you make an updated archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
