+++ report/study-flow-while-running.Rmd
@@ -41,6 +41,7 @@ root.path <- "/Volumes/DOS/daten/2016/"
 
 # Set directories
 processed.data.directory <- "processed-data/"
+feature.directory <- "features/"
 activity.directory <- "running/"
 
 ```
@@ -48,17 +49,12 @@ activity.directory <- "running/"
 ```{r load-data, include=FALSE}
 
 # Load user directories
-user.directories <- paste(list.dirs(path = paste(root.path, processed.data.directory, sep = ""), full.names = F, recursive = F), "/", sep = "")
+user.directories <- paste(list.dirs(path = paste(root.path, processed.data.directory, activity.directory, sep = ""), full.names = F, recursive = F), "/", sep = "")
 user.count <- length(user.directories)
-
-# Set directories
-feature.directory <- "features/"
+user.efforts <- c(6, 8, 6, 5, 5, 8, 7, 8, 7, 7, 6, 7, 7, 8, 6, 5, 7, 6, 7, 8, 5, 6, 4, 7, 6, 6, 7, 7, 9, 7, 5, 7, 3, 6, 5, 7)
 
 # Collect data
 data.1 <- data.frame()
-
-#sub.workload <- c(1,2,3,1,1,3,2,2,3,3,2,3,3,3,1,1,3,1,3,2,3,1,2,1,2,1,3,1,3,1,1,1,1,1,3,1)
-self.workload <- c(6,8,6,5,5,8,7,8,7,7,6,7,7,8,6,5,7,6,7,8,5,6,4,7,6,6,7,7,9,7,5,7,3,6,5,7)
 
 for (u in 1:user.count) {
   
@@ -71,18 +67,18 @@ for (u in 1:user.count) {
   fss.features <- read.csv(fss.feature.path)
   rm(fss.feature.path)
   
-  # Get HRV features 
+  # Set begin of the timestamp before the end of the data
   min.before.end <- 15
+  
+  # Get HRV features 
   source("code-snippets/get-hrv-features.R")
   
   # Left leg
   # Get stride features
-  min.before.end <- 15
   kinematic.data.file.name <- "imu-rn42-bc98"
   source("code-snippets/get-stride-features.R")
     
   # Get jerk cost features
-  min.before.end <- 15
   kinematic.data.file.name <- "imu-rn42-bc98"
   source("code-snippets/get-jc-features.R")
   mean.ms.interval.left = stride.features[, 1]
@@ -90,56 +86,63 @@ for (u in 1:user.count) {
 
   # Right leg
   # Get stride features
-  min.before.end <- 15
   kinematic.data.file.name <- "imu-rn42-3b70"
   source("code-snippets/get-stride-features.R")
     
   # Get jerk cost features
-  min.before.end <- 15
   kinematic.data.file.name <- "imu-rn42-3b70"
   source("code-snippets/get-jc-features.R")
   mean.ms.interval.right = stride.features[, 1]
   mean.jerk.cost.right = jc.features[, 1]
     
   # Get cardio-locomotor phase synchronization features
-  min.before.end <- 15
   source("code-snippets/get-cls-features.R") 
   
   # Get
   age <- trunc(as.numeric(as.Date(fss.features[, 17]) - as.Date(fss.features[, 9])) / -365)
   
   # Create data frame
-  data.1 <- rbind(data.1, data.frame(subject = subjects, fss.features[, c(1:8)], hrv.features[, 1:2], cls.features[, 1:3], mean.ms.interval.left, mean.ms.interval.right, mean.jerk.cost.left, mean.jerk.cost.right, age))
[…]
