+++ report/study-flow-while-running.Rmd
@@ -102,7 +102,7 @@ for (u in 1:user.count) {
   age <- trunc(as.numeric(as.Date(fss.features[, 17]) - as.Date(fss.features[, 9])) / -365)
   
   # Create data frame
-  data.1 <- rbind(data.1, data.frame(subject = subjects, fss.features[, c(1:7)], user.effort = user.efforts[u], hrv.features[, 1:2], cls.features[, 1:3], stride.freq.left = 60000 / mean.ms.interval.left, stride.freq.rigth = 60000 / mean.ms.interval.right, mean.jerk.cost.left / 10^3, mean.jerk.cost.right / 10^3, age))
+  data.1 <- rbind(data.1, data.frame(subject = subjects, fss.features[, c(1:7)], user.effort = user.efforts[u], hrv.features[, 1:2], cls.features[, 1:3], mean.stride.freq.left = 60000 / mean.ms.interval.left, mean.stride.freq.right = 60000 / mean.ms.interval.right, mean.jerk.cost.left / 10^3, mean.jerk.cost.right / 10^3, age))
 }
 
 rownames(data.1) <- seq(length = nrow(data.1))
@@ -113,39 +113,48 @@ rm(u, user.directory, user.efforts, subjects, fss.features, hrv.features, cls.fe
 ```{r view, echo=FALSE, warning=FALSE, eval=FALSE}
 
 # Remove users
-
 data.2 <- cbind(data.1, user.directories)
 data.3 <- data.2
-data.3 <- data.3[-c(28, 29, 33, 36), ] #10, 15, 26, 32), ]
+data.3 <- data.3[-c(28, 29, 33, 36, 10, 15, 26, 32), ]
 
 
 physiological.state <- data.3$direction
 physiological.state[data.3$mean.nsei > .1] <- 2
 
-data.3 <- data.3[, c(2:4,6,10,13,16,18,19)]
-plot(data.3, col = physiological.state)
-
+plot(data.2, col = physiological.state)
 apa.cor.table(data.2, show.conf.interval = FALSE)
+
+data.3 <- data.3[, c(2:4,10,13,16,18)]
+
 
 feature <- 3
 
-boxplot(data.3[, feature] ~ physiological.state)
+boxplot(data.3[, feature] ~ factor(physiological.state, labels = c("Herz dominiert", "Gleichgewicht", "Schritt dominiert")), ylab = "Absorbiertheit")
 
-hist(data.3[physiological.state == 1, feature])
-length(data.3[physiological.state == 1, feature])
-shapiro.test(data.3[physiological.state == 1, feature])
-hist(data.3[physiological.state == 2, feature])
-length(data.3[physiological.state == 2, feature])
-shapiro.test(data.3[physiological.state == 2, feature])
-hist(data.3[physiological.state == 3, feature])
-length(data.3[physiological.state == 3, feature])
-shapiro.test(data.3[physiological.state == 3, feature])
+# hist(data.3[physiological.state == 1, feature])
+# length(data.3[physiological.state == 1, feature])
+# shapiro.test(data.3[physiological.state == 1, feature])
+# hist(data.3[physiological.state == 2, feature])
+# length(data.3[physiological.state == 2, feature])
+# shapiro.test(data.3[physiological.state == 2, feature])
+# hist(data.3[physiological.state == 3, feature])
+# length(data.3[physiological.state == 3, feature])
+# shapiro.test(data.3[physiological.state == 3, feature])
 
 t.test(data.3[physiological.state == 1, feature], data.3[physiological.state == 2, feature])
 t.test(data.3[physiological.state == 2, feature], data.3[physiological.state == 3, feature])
 
-5 minuten abschnitte
[…]
