### 2016-09-10 #41 @arfon
@sbogutzky thanks for updating the DOI here. @davclark many thanks for the thorough review.

@sbogutzky your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00041 🎉 🚀 💥
