### 2024-03-22 #41 @xuanxu
Hey @arfon, the value for the archive is incorrect, it should be set to `10.5281/zenodo.61965` adn then reaccept the paper. Can you take care of this?

As [reported here](https://github.com/openjournals/joss/issues/1324) by @sdruskat
